refactor: 필요없는 Bootpay 라이브러리 삭제
- 코드로 갖고 있을 필요 없었다
- 개발 문서도 코드 넣으라는 말 없음
