--- src/main/java/asc/portfolio/ascSb/bootpay/domain/DevMode.java
@@ -1,8 +1,0 @@
-package asc.portfolio.ascSb.bootpay.domain;
-
-class DevMode {
-    public static final String DEVELOPMENT = "https://dev-api.bootpay.co.kr/";
-    public static final String TEST = "https://test-api.bootpay.co.kr/";
-    public static final String STAGE = "https://stage-api.bootpay.co.kr/";
-    public static final String PRODUCTION = "https://api.bootpay.co.kr/";
-}
