--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/Item.java
@@ -1,11 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-public class Item {
-    public String item_name; // 상품명
-    public int qty; // 수량
-    public String unique; // 상품 고유키
-    public long price; // 상품단가
-    public String cat1; // 카테고리 상
-    public String cat2; // 카테고리 중
-    public String cat3; // 카테고리 하
-}
