--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/LMS.java
@@ -1,12 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Deprecated
-public class LMS {
-    public String sp;
-    public List<String> rps = new ArrayList();
-    public String msg;
-    public String sj;
-}
