--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/RefundData.java
@@ -1,7 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-public class RefundData {
-    public String account; // 환불받을 계좌번호
-    public String accountholder; // 환불받을 계좌주
-    public String bankcode; //  은행 코드
-}
