--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/SMS.java
@@ -1,11 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Deprecated
-public class SMS {
-    public String sp;
-    public List<String> rps = new ArrayList();
-    public String msg;
-}
