--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/Submit.java
@@ -1,5 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-public class Submit {
-    public String receiptId;
-}
