--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/Token.java
@@ -1,7 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-public class Token {
-    public String application_id; // 부트페이 관리자에서 확인 가능한 rest application id
-    public String private_key; // 부트페이 관리자에서 확인 가능한 private key
-}
-
