--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/request/UserToken.java
@@ -1,11 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.request;
-
-public class UserToken {
-    public String userId; // 개발사에서 관리하는 회원 고유 id
-    public String email; // 회원 email
-    public String name; // 회원명
-    public int gender; // 0:여자, 1:남자
-    public String birth; // 생일 901004
-    public String phone; //01012341234
-}
-
