--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/ResDefault.java
@@ -1,17 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response;
-
-import com.google.gson.Gson;
-
-/**
- * Created by ehowlsla on 2018. 5. 29..
- */
-public class ResDefault<T> {
-    public int status;
-    public int code;
-    public String message;
-    public T data;
-
-    public String toJson() {
-        return new Gson().toJson(this);
-    }
-}
