--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/BillingKeyCardData.java
@@ -1,8 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class BillingKeyCardData {
-    public String card_code;
-    public String card_name;
-    public String card_no;
-    public String card_cl;
-}
