--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/CancelData.java
@@ -1,12 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class CancelData {
-    public String receipt_id;
-    public double request_cancel_price;
-    public double remain_price;
-    public double remain_tax_free;
-    public double cancelled_price;
-    public double cancelled_tax_free;
-    public String revoked_at;
-    public String tid;
-}
