--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/EasyUserTokenData.java
@@ -1,7 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class EasyUserTokenData {
-    public String user_token;
-    public long expired_unixtime;
-    public String expired_localtime;
-}
