--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/PaymentData.java
@@ -1,24 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class PaymentData {
-    public String username;
-    public String phone;
-    public String birth;
-    public String gender;
-    public String unique;
-    public String di;
-    public String n;
-
-    public int p;
-    public String tid;
-    public String pg;
-    public String pm;
-
-    public String pg_a;
-    public String pm_a;
-    public String o_id;
-    public String p_at;
-
-    public int s;
-    public int g;
-}
