--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/SubscribeBillingData.java
@@ -1,24 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class SubscribeBillingData {
-    public String receipt_id;
-    public double price;
-    public String card_no;
-    public String card_code;
-    public String card_name;
-
-    public String params;
-    public String item_name;
-    public String order_id;
-    public String url;
-    public String payment_name;
-
-    public String pg_name;
-    public String pg;
-
-    public String method;
-    public String method_name;
-    public String requested_at;
-    public String purchased_at;
-    public int status;
-}
