--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/SubscribeBillingReserveData.java
@@ -1,12 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class SubscribeBillingReserveData {
-    public String reserve_id;
-    public String billing_key;
-    public String item_name;
-    public double price;
-    public int purchase_count;
-    public int purchase_limit;
-    public long execute_at;
-    public int status;
-}
