--- src/main/java/asc/portfolio/ascSb/bootpay/dto/model/response/data/TokenData.java
@@ -1,7 +1,0 @@
-package asc.portfolio.ascSb.bootpay.dto.model.response.data;
-
-public class TokenData {
-    public String token;
-    public long server_time;
-    public long expired_at;
-}
